refactor: rename PressureChart -> PressureResponseChart, BandsChart -> PressureBandsChart
Renames the two chart components for clarity:
  - PressureChart      -> PressureResponseChart (Chart.js pressure-response line chart)
  - BandsChart         -> PressureBandsChart    (Piaf/Pmax banded marker chart)

Updates every import, usage, comment, doc, the e2e comment, and the
screenshot-capture labels. Internal CSS classes (.bands-chart, .chart-wrap)
and the exported BandMarker type are unchanged. Also bumps the data-repo
pointer for the matching comment update in interpolate.ts.
